Remove extra memory usage in batch processing.
Output of previous process_function call was not deleted yet when process_function is called next time and returns new output.
If output use much memory (say, in segmentation tasks), it results in 'out of memory' error for batch size smaller then it could be if previous output is released before new process_function call.

Update: possibly some specific process_functions may need previous output. So it is better not to release it at engine level but in process_function.