Update __init__.py (#489)
Tiny fix to change `metrics={}` to `metrics=None` and using `None` as a sentinel value for an empty `dict`.

Mutable default arguments get weird because they are created once when the signature is declared and can be mutated inside of the function's body. It wasn't an issue in this function but if anyone used this as an example to write their own evaluator, they may stumble into some tricky bugs if they start mutating `metrics`.