Kais fix for `create_supervised` bug
Now the model will always be moved before processing the data and
calculating the output. This is similar to the process how the
data is preprocessed/moved. The `.to()` operation becomes a
No-Op if the model is already on the correct device.
This gives the user more options and works, like documentation hints.
(Model will be moved correctly for run)