Changed fix to vfdev-5 proposal
There is a performance penalty for using `model.to()`
at each iteration, even if the model is on the same device.
The model is now moved when the engine starts and the
`Events.STARTED` Even is triggered.

The documentation was rewritten, to make it clear that the model
is *only* moved with the start of the engine.

The test cases covering the model move between iterations
are now removed.