Create supverised fix (#835)
* Added create_supervised tests

Now the current master is failing,
as the model is moved too early in the `create_supervised` function

* Kais fix for `create_supervised` bug

Now the model will always be moved before processing the data and
calculating the output. This is similar to the process how the
data is preprocessed/moved. The `.to()` operation becomes a
No-Op if the model is already on the correct device.
This gives the user more options and works, like documentation hints.
(Model will be moved correctly for run)

* Added `State` to __all__ in engine/__init__

Removed in previous commit, as it wasn't used, but was just forgotten
in the __all__ list.

* Changed fix to vfdev-5 proposal

There is a performance penalty for using `model.to()`
at each iteration, even if the model is on the same device.
The model is now moved when the engine starts and the
`Events.STARTED` Even is triggered.

The documentation was rewritten, to make it clear that the model
is *only* moved with the start of the engine.

The test cases covering the model move between iterations
are now removed.

Co-authored-by: vfdev <vfdev.5@gmail.com>